requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
anthropic>=0.39.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
gspread>=5.12.0
python-telegram-bot>=21.0
deepgram-sdk>=3.0.0
